AutoLogin Implementation
-AutoLogin Implementation
-LoginScreen AutoLogin Checkbox
-AsyncStorage API Construction
-Login Data Handling 간소화
